+++ CHANGELOG.md
@@ -66,6 +66,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - customer_type attribute to IssuingProduct resource
 - Rule sub-resource and rules, debtor_workspace_id and nominal_interest attributes to CreditNote resource
 - BrcodePreview resource
+- IssuingTokenDesign resource
 ### Changed
 - IssuingBin resource to IssuingProduct
 - settlement parameter to funding_type and client parameter to holder_type of IssuingProduct